requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.5.0
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
edge-tts==7.0.0
email_validator==2.2.0
faster-whisper==1.1.1
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
boto3
gitdb==4.0.12
GitPython==3.1.44
groovy==0.1.2
groq==0.22.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
janus==1.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
miniaudio==1.61
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
narwhals==1.32.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
onnxruntime==1.21.0
openai==1.69.0
openai-whisper==20240930
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.55
pillow==11.1.0
propcache==0.3.1
protobuf==4.25.6
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyht==0.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.0
rpds-py==0.24.0
ruff==0.11.2
s3transfer==0.11.4
safehttpx==0.1.6
semantic-version==2.10.0
sentry-sdk==2.24.1
setuptools==78.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.7
srt==3.5.3
starlette==0.46.1
streamlit==1.44.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.4
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
